+++ ui-src/js/modal.js
@@ -2,30 +2,46 @@ angular
     .module('nzbhydraApp')
     .factory('ModalService', ModalService);
 
-function ModalService($uibModal) {
+function ModalService($uibModal, $q) {
     
     return {
-        open: openModal
+        open: open
     };
     
-    function openModal(headline, message, ok, cancel) {
+    function open(headline, message, params) {
+        //params example:
+        /*
+        var p =
+        {
+            yes: {
+                text: "Yes",    //default: Ok
+                onYes: function() {}
+            },
+            no: {               //default: Empty
+                text: "No",
+                onNo: function () {
+                }
+            },
+            cancel: {           
+                text: "Cancel", //default: Cancel
+                onCancel: function () {
+                }
+            }
+        };
+        */
         var modalInstance = $uibModal.open({
             templateUrl: 'static/html/modal.html',
             controller: 'ModalInstanceCtrl',
             size: 'md',
             resolve: {
                 headline: function () {
-                    return headline
+                    return headline;
                 },
-                message: function(){ return message},
-                ok: function() {
-                    return ok;
+                message: function(){ 
+                    return message;
                 },
-                cancel: function() {
-                    return cancel;
-                },
-                showCancel: function() {
-                    return angular.isDefined(cancel);
+                params: function() {
+                    return params;
                 }
             }
         });
@@ -33,8 +49,7 @@ function ModalService($uibModal) {
         modalInstance.result.then(function() {
             
         }, function() {
-            if (angular.isDefined(cancel))
-            cancel();
+            
         });
     }
     
@@ -44,23 +59,50 @@ angular
     .module('nzbhydraApp')
     .controller('ModalInstanceCtrl', ModalInstanceCtrl);
 
-function ModalInstanceCtrl($scope, $uibModalInstance, headline, message, ok, cancel, showCancel) {
+function ModalInstanceCtrl($scope, $uibModalInstance, headline, message, params) {
 
     $scope.message = message;
     $scope.headline = headline;
-    $scope.showCancel = showCancel;
+    $scope.params = params;
+    $scope.showCancel = angular.isDefined(params.cancel);
+    $scope.showNo = angular.isDefined(params.no);
 
-    $scope.ok = function () {
+    if (angular.isDefined(params.yes) && angular.isUndefined(params.yes.text)) {
+        params.yes.text = "Yes";
+    }
+    
+    if (angular.isDefined(params.no) && angular.isUndefined(params.no.text)) {
+        params.no.text = "No";
+    }
+    
+    if (angular.isDefined(params.cancel) && angular.isUndefined(params.cancel.text)) {
+        params.cancel.text = "Cancel";
+    }
+
+    $scope.yes = function () {
         $uibModalInstance.close();
-        if(!angular.isUndefined(ok)) {
-            ok();
+        if(angular.isDefined(params.yes) && angular.isDefined(params.yes.onYes)) {
+            params.yes.onYes();
+        }
+    };
+
+    $scope.no = function () {
+        $uibModalInstance.close();
+        if (angular.isDefined(params.no) && angular.isDefined(params.no.onNo)) {
+            params.no.onNo();
         }
     };
 
     $scope.cancel = function () {
         $uibModalInstance.dismiss();
-        if (!angular.isUndefined(cancel)) {
-            cancel();
+        if (angular.isDefined(params.cancel) && angular.isDefined(params.cancel.onCancel)) {
+            params.cancel.onCancel();
         }
     };
+
+    $scope.$on("modal.closing", function (targetScope, reason, c) {
+        if (reason == "backdrop click") {
+            $scope.cancel();
+        }
+    });
 }